+++ services/application/models/home_model.php
@@ -391,13 +391,15 @@ class home_model extends CI_Model
       return null;
     }
   }
-  public function checkUserExists($post)
+  public function checkUserExists(array $post)
   {
     $requestType = $post["requestType"];
-    $query = $this->db
-      ->where(["user_name" => strtolower($post["username"])])
-      ->or_where("user_email", strtolower($post["email"]))
-      ->get("users");
+    $query = $this->db->where(["user_name" => strtolower($post["username"])])->get("users");
+    /**
+     * Important:
+     * Only unique usernames are validated
+     * Emails can be multiple, for multiple accounts
+     */
     $numRows = $query->num_rows();
     $flag = true;
     if ($requestType === "Create" && $numRows > 0) {
@@ -469,10 +471,10 @@ class home_model extends CI_Model
       return false;
     }
   }
-  public function validateOtpTime($post)
+  public function validateOtpTime(array $post)
   {
     $this->db->where([
-      "user_id" => $post["id"],
+      "user_name" => $post["id"],
       "user_otp" => $post["otp"],
       "user_otp_expiry >" => time(),
     ]);
@@ -483,9 +485,9 @@ class home_model extends CI_Model
       return false;
     }
   }
-  public function resetUpdate($userId, $resetPassword)
+  public function resetUpdate(string $userId, string $resetPassword)
   {
-    $this->db->where("user_id", $userId);
+    $this->db->where("user_name", $userId);
     $this->db->update("users", ["user_password" => md5($resetPassword)]);
     if ($this->db->affected_rows() > 0) {
       return true;